Use rainbow presenter, don't monkey-patch String.
Some gems monkey-patch String with a (possibly private) `color` method, which can cause this gem to fail.

At least one other guard plugin I'm using relies on a gem that does this.  This change uses the rainbow presenter API to colorize messages, so this plugin will behave as a good citizen and not add a monkey-patch to String.  As a bonus, it will be more resistant to breakage from bad-behaving gems.